Added Exceptions, added switch in main method, added method signatures in main method, added recursive category search in MainController, implemented tryPrintTrolls, printCategoryTree

+++ shop-app/src/main/java/com/unileipzig/shop/repository/ProductRepository.java
@@ -80,7 +80,7 @@ public class ProductRepository {
         }
     }
 
-    public List<Product> getProductsByCategoryId(int catID) {
+    public List<Product> getProductsByCategoryId(long catID) {
         Session session = HibernateConnector.getSession();
         Query<Product> query = session.createQuery("SELECT p FROM Category c INNER JOIN c.products p " +
                 "WHERE c.id = :catID", Product.class);
